Add scaffolding for sessions and admin
Allow admin users to login and logout.
After logging in with a valid namd and password combination, redirect
the admin user to the admin page.

The admin page displays the current time and how many orders we have.

If the name/password combination is invalid, alert the user and prompt
them to login again.