Only allow authenticated users to access admin pages
Only authenticated users should be able to access store admin pages.
Shoppers should still be able to access store pages so that they can
view the products, add products to their carts, and place orders.

For tests, log the user in so that they are not redirected to the admin
page. When testing deleting a customer, be authenticated with a different
user than the one being deleted.